+++ standard/packages/preact/src/daygrid/components/DayGridLayoutNormal.tsx
@@ -86,7 +86,7 @@ export class DayGridLayoutNormal extends BaseComponent<DayGridLayoutNormalProps,
               borderlessX,
               borderlessTop,
               borderlessBottom,
-              multiMonthColumnCount: 0,
+              multiMonthColumns: 0,
             }),
             classNames.printHeader, // either flexCol or table-header-group
             tableHeaderSticky && classNames.tableHeaderSticky,
@@ -110,7 +110,7 @@ export class DayGridLayoutNormal extends BaseComponent<DayGridLayoutNormalProps,
             <div
               className={generateClassName(options.dayHeaderDividerClass, {
                 isSticky: tableHeaderSticky,
-                multiMonthColumnCount: 0,
+                multiMonthColumns: 0,
                 options: { allDaySlot: Boolean(options.allDaySlot) },
               })}
             />
@@ -123,7 +123,7 @@ export class DayGridLayoutNormal extends BaseComponent<DayGridLayoutNormalProps,
               borderlessX,
               borderlessTop,
               borderlessBottom,
-              multiMonthColumnCount: 0,
+              multiMonthColumns: 0,
             }),
             // HACK for Safari. Can't do break-inside:avoid with flexbox items, likely b/c it's not standard:
             // https://stackoverflow.com/a/60256345