+++ standard/packages/preact/src/themes/classic/index.tsx
@@ -342,11 +342,11 @@ export default {
     --------------------------------------------------------------------------------------------- */
 
     singleMonthClass: (info) => joinClassNames(
-      info.multiMonthColumnCount > 1 && 'm-4',
-      (info.multiMonthColumnCount === 1 && !info.isLast) && 'border-b border-(--fc-classic-border)',
+      info.multiMonthColumns > 1 && 'm-4',
+      (info.multiMonthColumns === 1 && !info.isLast) && 'border-b border-(--fc-classic-border)',
     ),
     singleMonthHeaderClass: (info) => joinClassNames(
-      info.multiMonthColumnCount > 1
+      info.multiMonthColumns > 1
         ? 'pb-4'
         : 'py-2 border-b border-(--fc-classic-border) bg-(--fc-classic-background)',
       'items-center',
@@ -423,7 +423,7 @@ export default {
       ...dayRowCommonClasses,
       dayCellBottomClass: 'min-h-px',
       tableClass: (info) => joinClassNames(
-        info.multiMonthColumnCount > 1 && 'border-(--fc-classic-border) border',
+        info.multiMonthColumns > 1 && 'border-(--fc-classic-border) border',
       ),
     },
     timeGrid: {