+++ standard/packages/preact/src/timegrid/components/TimeGridLayout.tsx
@@ -168,7 +168,7 @@ export class TimeGridLayout extends BaseComponent<TimeGridLayoutProps> {
             borderlessX,
             borderlessTop,
             borderlessBottom,
-            multiMonthColumnCount: 0,
+            multiMonthColumns: 0,
           }),
           // we don't do classNames.printRoot/classNames.printHeader here because works poorly with print:
           // - Firefox >85ish CAN have flexboxes within it, but those cannot do absolute positioning