+++ standard/packages/preact/src/timegrid/components/TimeGridLayoutPannable.tsx
@@ -206,7 +206,7 @@ export class TimeGridLayoutPannable extends BaseComponent<TimeGridLayoutPannable
                 borderlessX,
                 borderlessTop,
                 borderlessBottom,
-                multiMonthColumnCount: 0,
+                multiMonthColumns: 0,
               }),
               // see note in TimeGridLayout about why we don't do classNames.printHeader
               classNames.flexCol,
@@ -310,7 +310,7 @@ export class TimeGridLayoutPannable extends BaseComponent<TimeGridLayoutPannable
             <div
               className={generateClassName(options.dayHeaderDividerClass, {
                 isSticky: tableHeaderSticky,
-                multiMonthColumnCount: 0,
+                multiMonthColumns: 0,
                 options: { allDaySlot: Boolean(options.allDaySlot) },
               })}
             />
@@ -323,7 +323,7 @@ export class TimeGridLayoutPannable extends BaseComponent<TimeGridLayoutPannable
               borderlessX,
               borderlessTop,
               borderlessBottom,
-              multiMonthColumnCount: 0,
+              multiMonthColumns: 0,
             }),
             classNames.flexCol,
             verticalScrolling && classNames.liquid,