+++ theming/theme-classic-tailwind/src/options-event-calendar.ts
@@ -374,13 +374,13 @@ export function createEventCalendarOptions(params: EventCalendarOptionParams): {
       ------------------------------------------------------------------------------------------- */
 
       singleMonthClass: (info) => joinClassNames(
-        info.multiMonthColumnCount > 1 && 'm-4',
-        (info.multiMonthColumnCount === 1 && !info.isLast) &&
+        info.multiMonthColumns > 1 && 'm-4',
+        (info.multiMonthColumns === 1 && !info.isLast) &&
           `${params.borderColorClass} border-b`,
       ),
 
       singleMonthHeaderClass: (info) => joinClassNames(
-        info.multiMonthColumnCount > 1
+        info.multiMonthColumns > 1
           ? 'pb-4'
           : `py-2 border-b ${params.borderColorClass} ${params.bgClass}`,
         'items-center', // h-align
@@ -430,7 +430,7 @@ export function createEventCalendarOptions(params: EventCalendarOptionParams): {
         dayCellBottomClass: 'min-h-px',
 
         tableClass: (info) => joinClassNames(
-          info.multiMonthColumnCount > 1 && `${params.borderColorClass} border`
+          info.multiMonthColumns > 1 && `${params.borderColorClass} border`
         )
       },
       timeGrid: {