+++ theming/ui-default-react-tailwind/src/theme-pulse/_compiled/event-calendar.tsx
@@ -560,11 +560,11 @@ export function EventCalendar({
       ------------------------------------------------------------------------------------------- */
 
       singleMonthClass={(info) => joinClassNames(
-        info.multiMonthColumnCount > 1 && 'm-3',
-        (info.multiMonthColumnCount === 1 && !info.isLast) && 'border-b border-(--fc-pulse-border)',
+        info.multiMonthColumns > 1 && 'm-3',
+        (info.multiMonthColumns === 1 && !info.isLast) && 'border-b border-(--fc-pulse-border)',
       )}
       singleMonthHeaderClass={(info) => joinClassNames(
-        info.multiMonthColumnCount > 1
+        info.multiMonthColumns > 1
           ? 'pb-2'
           : 'py-1 border-b border-(--fc-pulse-border) bg-(--fc-pulse-background)',
         'items-center',
@@ -617,12 +617,12 @@ export function EventCalendar({
         multiMonth: {
           ...dayRowCommonClasses,
           viewClass: 'bg-(--fc-pulse-faint)',
-          tableHeaderClass: (info) => joinClassNames(info.multiMonthColumnCount === 1 && 'bg-(--fc-pulse-background)'),
+          tableHeaderClass: (info) => joinClassNames(info.multiMonthColumns === 1 && 'bg-(--fc-pulse-background)'),
           tableBodyClass: (info) => joinClassNames(
-            info.multiMonthColumnCount > 1 && 'border border-(--fc-pulse-border) rounded-sm overflow-hidden'
+            info.multiMonthColumns > 1 && 'border border-(--fc-pulse-border) rounded-sm overflow-hidden'
           ),
           dayHeaderAlign: (info) => info.inPopover ? 'start' : info.isNarrow ? 'center' : 'end',
-          dayHeaderDividerClass: (info) => joinClassNames(info.multiMonthColumnCount === 1 && 'border-b border-(--fc-pulse-border)'),
+          dayHeaderDividerClass: (info) => joinClassNames(info.multiMonthColumns === 1 && 'border-b border-(--fc-pulse-border)'),
           dayCellBottomClass: getShortDayCellBottomClass,
           ...userViews?.multiMonth,
         },