+++ standard/CHANGELOG.md
@@ -7,6 +7,8 @@
 - FIX: In React/Preact, expose the CalendarRef type
 - FEATURE: Warn on unknown options
 - FEATURE: Warn on invalid class-name inputs
+- BREAKING (TODO: update in docs)
+  - renamed multiMonthColumnCount -> multiMonthColumns
 
 
 ## v7.0.0-rc.2